pyproject.toml
[project]
name = "okta-python-dpop-example"
version = "1.0.0"
description = "Python implementation of Okta DPoP (Demonstrating Proof of Possession) example"
authors = [
    {name = "Developer", email = "dev@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "PyJWT>=2.8.0",
    "cryptography>=41.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "jwcrypto>=1.5.6",
    "pytest>=8.3.5",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
